requirements.txt
PyYAML
paho-mqtt
asyncio
requests

